+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Please notice we'll use semantic versioning as of 1.0.0.
 
 ## 0.5.0 - Unreleased
 
+### Added
+
+* Added an `encrypt` command to encrypt backup files using GPG.
+* Added a `decrypt` CLI command to decrypt backup files using GPG.
+
 ## 0.4.0 - 2015-02-25
 
 ### Added